+++ src/cone/app/browser/contents.py
@@ -1,3 +1,4 @@
+from cone.app.browser import RelatedViewProvider
 from cone.app.browser import render_main_template
 from cone.app.browser.actions import ActionDelete
 from cone.app.browser.actions import ActionEdit
@@ -14,6 +15,7 @@ from cone.app.interfaces import IWorkflowState
 from cone.tile import Tile
 from cone.tile import tile
 from node.utils import instance_property
+from plumber import plumbing
 from pyramid.i18n import TranslationStringFactory
 from pyramid.security import has_permission
 from pyramid.view import view_config
@@ -196,9 +198,11 @@ class ContentsTile(Table):
 
 
 @tile(name='listing', path='templates/listing.pt', permission='list')
+@plumbing(RelatedViewProvider)
 class ListingTile(Tile):
     """Tile rendering the listing.
     """
+    related_view = 'listing'
 
 
 @view_config(name='listing', permission='list')